+++ hazimp/console.py
@@ -43,7 +43,7 @@ def build_cmd_parser():
 
     parser.add_argument('-v', '--version',
                         action='version',
-                        version="%(prog)s 0.1")
+                        version="%(prog)s 1.1")
     return parser