+++ tests/test_misc.py
@@ -51,7 +51,8 @@ from hazimp.misc import (csv2dict, get_required_args, sorted_dict_values,
                          get_temporary_directory, s3_path_segments_from_vsis3,
                          create_temp_file_path_for_s3,
                          download_from_s3, upload_to_s3_if_applicable,
-                         download_file_from_s3_if_needed, mod_file_list, permutate_att_values, get_git_commit)
+                         download_file_from_s3_if_needed, mod_file_list,
+                         permutate_att_values, get_git_commit)
 from tests import CWD
 
 
@@ -224,13 +225,13 @@ class TestMisc(unittest.TestCase):
             permutate_att_values(data_frame, 'x', 'y')
 
     def test_get_git_commit(self):
-        self.assertEqual((ANY, ANY, ANY), get_git_commit())
+        self.assertEqual((ANY, ANY, ANY, ANY), get_git_commit())
 
     @patch.object(Repo, 'commit')
     def test_get_git_commit_without_git(self, mock):
         mock.side_effect = InvalidGitRepositoryError()
 
-        self.assertEqual(('unknown', '', ANY), get_git_commit())
+        self.assertEqual(('unknown', '', ANY, 'unknown'), get_git_commit())
 
     def test_get_s3_client(self):
         s3_client = get_s3_client()