+++ docs/history.rst
@@ -1,7 +1,24 @@
-.. _history:
-
 History
 =======
+
+1.1 (2021-07-XX)
+----------------
+
+* Add permutation of exposure attributes
+
+1.0.2 (2021-05-05)
+------------------
+
+* Minor updates to complete development of aggregation functions.
+* Supports multiple output formats
+* Aggregated features can now be categorised (labelled) based on selected structural loss values (e.g. mean, max, min)
+
+
+1.0.1 (2021-03-17)
+------------------
+
+* Hotfix update to include a new suite of wind vulnerability functions
+
 
 1.0.0 (2021-03-12)
 ------------------